작업사항(21/10/02)
1. Resource
- Animation 클래스 완료

2. Component
- Camera 클래스 구현 중
- Transform 클래스 구현 중

3. TODO
- Camera, Transform 클래스 구현 마무리

+++ Project/DX11_2D_GameEngine_Lib/InputLayout.cpp
@@ -101,7 +101,7 @@ const bool InputLayout::Create(ID3DBlob* blob)
     hResult = device->CreateInputLayout
 	(
 		input_layout_element_descs.data(),
-		input_layout_element_descs.size(),
+		static_cast<UINT>(input_layout_element_descs.size()),
 		blob->GetBufferPointer(),
 		blob->GetBufferSize(),
 		m_p_input_layout.GetAddressOf()
